refactor, dev: поправлено название модуля
